requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.4
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
schedule==1.2.0
line-bot-sdk==3.5.0
lxml==4.9.3